# OpenApi/MysiePysieService.yaml
openapi: 3.0.0
info:
  title: MysiePysie Service
  description: |
    Mysie Pysie Service for UZ Project with collaboration with Metapack
  contact:
    email: bartosz_musielak@hotmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
- url: https://virtserver.swaggerhub.com/MysiePysie/MysiePysie/0.0.1
  description: SwaggerHub API Auto Mocking
- url: localhost:8080
paths:
...
  /teachers:
    get:
      tags:
      - teachers
      summary: Returns all existing teachers
      operationId: getTeachers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        "404":
          description: Teachers not found
        "500":
          description: Server error
    put:
      tags:
      - teachers
      summary: Update an existing teacher
      operationId: updateTeacher
      requestBody:
        $ref: '#/components/requestBodies/Teacher'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Teacher not found
        "405":
          description: Validation exception
    post:
      tags:
      - teachers
      summary: Add a new teacher to the database
      operationId: addTeacher
      requestBody:
        $ref: '#/components/requestBodies/Teacher'
      responses:
        "405":
          description: Invalid input
...
  /teachers/{teacherId}:
    get:
      tags:
      - teachers
      summary: Find teacher by ID
      description: Returens a single teacher
      operationId: getTeacherById
      parameters:
      - name: teacherId
        in: path
        description: ID of teacher to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
            application/xml:
              schema:
                $ref: '#/components/schemas/Teacher'
        "400":
          description: Invalid ID supplied
        "404":
          description: Teacher not found
    post:
      tags:
      - teachers
      summary: Updates a teacher in the database with form data
      operationId: updateTeacherWithForm
      parameters:
      - name: teacherId
        in: path
        description: ID of teacher that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/body_1'
      responses:
        "405":
          description: Invalid input
    delete:
      tags:
      - teachers
      summary: Deletes a teacher
      operationId: deleteTeacher
      parameters:
      - name: teacherId
        in: path
        description: Teacher id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Teacher not found
...
  /user/login:
    get:
      tags:
      - user
      summary: Logs user into the system
      operationId: loginUser
      parameters:
      - name: username
        in: query
        description: The user name for login
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: password
        in: query
        description: The password for login in clear text
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              style: simple
              explode: false
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              style: simple
              explode: false
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
        "400":
          description: Invalid username/password supplied
...
  /user/{username}:
    get:
      tags:
      - user
      summary: Get user by user name
      operationId: getUserByName
      parameters:
      - name: username
        in: path
        description: The name that needs to be fetched. Use user1 for testing.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
    put:
      tags:
      - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
      - name: username
        in: path
        description: name that need to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Updated user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        description: The name that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
components:
  schemas:
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
          format: int64
        forename:
          type: string
        surname:
          type: string
        age:
          type: integer
          format: int64
      xml:
        name: Teacher
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: integer
          description: User Status
          format: int32
      xml:
        name: User
...
    body_1:
      type: object
      properties:
        forename:
          type: string
          description: Updated forename of a teacher
        surname:
          type: string
          description: Updated surname of a teacher
        subject:
          type: string
          description: Updated subject of a teacher
        age:
          type: integer
          description: Updated age of a teacher
          format: int64
        status:
          type: string
          description: Updated status of a teacher
...
  requestBodies:
...
    Teacher:
      description: Teacher object that needs to be added to the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Teacher'
        application/xml:
          schema:
            $ref: '#/components/schemas/Teacher'
      required: true
...